% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ids.R
\name{convert_ids}
\alias{convert_ids}
\title{Converts sample id's from one format to another using a data frame with alternate information.}
\usage{
convert_ids(df, id_data = NULL, from_col = "native_id",
  to_col = "unified_id", other_cols = NULL)
}
\arguments{
\item{df}{A data frame with a column named sample_id}

\item{id_data}{A data frame containing information for id conversion}

\item{from_col}{Define the name of the column matching the old}

\item{to_col}{Define the name of the column matching the new ids}

\item{other_cols}{Other columns to include in the output}
}
\value{
a data frame
}
\description{
Converts sample id's from one format to another using a data frame with alternate information.
}
\examples{
data('CCLEsmall', package='PharmacoGx')

dat1 <- gather.PharmacoSet(CCLEsmall, sample_ids=c('143B', '23132-87'), gene_ids=c('BRAF', 'EGFR'),
 data_types=c('rna', 'mutation'), gene_col=c('Symbol', 'Symbol'))

data('CancerCellLineIDs')
convert_ids(df=dat1, id_data=dplyr::filter(CancerCellLineIDs, id_type=='CCLE'),
from_col='alt_id', to_col='unified_id')
convert_ids(df=dat1, id_data=dplyr::filter(CancerCellLineIDs, id_type=='CCLE'),
from_col='alt_id', to_col='unified_id', other_cols='tissue')

}

