% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crush.R
\name{crush}
\alias{crush}
\title{Crush cell line id format}
\usage{
crush(x)
}
\arguments{
\item{x}{character vector of ids}
}
\value{
character vector of crushed ids
}
\description{
Helper function to remove punctuation and whitespace from cell line id's.  This helps convert id's between datasets.
}
\examples{
data("CCLEsmall", package='PharmacoGx')
cellNames(CCLEsmall)[1:20]
crush(cellNames(CCLEsmall)[1:20])
}

