% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.ExpressionSet.R
\name{gather.ExpressionSet}
\alias{gather.ExpressionSet}
\title{Convert an ExpressionSet object into tidy format}
\usage{
\method{gather}{ExpressionSet}(x, sample_ids = NULL, gene_ids = NULL,
  sample_col = "cellid", gene_col = "Symbol")
}
\arguments{
\item{x}{An ExpressionSet}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene_ids}{A vector of gene ids.  Default is NULL (don't filter on gene id)}

\item{sample_col}{Name of the column in the pData data frame to use for filtering on sample id}

\item{gene_col}{Name of the column in the rowData data frame to use for filtering on gene id}
}
\value{
a data frame in tall_df format
}
\description{
Convert an ExpressionSet object into tidy format
}
\examples{
data('CCLEsmall', package='PharmacoGx')
eset <- CCLEsmall@molecularProfiles$rna
gather.ExpressionSet(eset, sample_ids=c('143B', '23132-87'), gene_ids=c('BRAF', 'EGFR') )
gather.ExpressionSet(eset, sample_ids=c('143B', '23132-87') )
}

