% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.PharmacoGx.R
\name{gather.PharmacoGx}
\alias{gather.PharmacoGx}
\title{gather.PharmacoGx}
\usage{
gather.PharmacoGx(x, sample_ids = NULL, gene_ids = NULL, resp_ids = NULL,
  sample_col = "cellid", gene_col = c("Symbol", "gene_name"),
  resp_col = "ic50_published", data_types = c("rna", "mutation"),
  resp_df = NULL)
}
\arguments{
\item{x}{PharmacoGx object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene_ids}{A vector of gene ids.  Default is NULL (don't filter on gene id)}

\item{resp_ids}{A vector of response ids.  Default is NULL (don't filter on response id)}

\item{sample_col}{Name of the column in the pData data frame to use for filtering on sample id}

\item{gene_col}{Name of the column in the rowData data frame to use for filtering on gene id}

\item{resp_col}{Response variable to retrieve}

\item{data_types}{Named of the componentd of the PharmacoGx object to gather}

\item{resp_df}{Data frame to process response data from (rather than PharmcaGx object)}
}
\value{
a tibble
}
\description{
gather.PharmacoGx
}
\examples{
data('CCLEsmall', package='PharmacoGx')
gather.PharmacoGx(CCLEsmall, sample_ids=c('143B', '23132-87'), gene_ids=c('BRAF', 'EGFR'),
 data_types=c('rna', 'mutation'), gene_col=c('Symbol', 'Symbol'))
gather.PharmacoGx(CCLEsmall, sample_ids=c('CHL-1', 'SW1573'), gene_ids=c('BRAF', 'EGFR'),
 resp_ids=c("AEW541","Nilotinib","PHA-665752","lapatinib"),
 data_types=c('rna', 'mutation'), gene_col=c('Symbol', 'Symbol'), resp_col='ic50_published')



}

