% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMultiAssay_generics.R
\name{gather}
\alias{gather}
\title{gather}
\usage{
gather(x, ...)
}
\arguments{
\item{x}{an ExpressionSet or PharmacoSet object}

\item{...}{other arguments see \code{\link{gather.ExpressionSet}} and \code{\link{gather.PharmacoSet}}}
}
\description{
S3 generic for \code{\link{gather.ExpressionSet}} and \code{\link{gather.PharmacoSet}}
defaulting to \code{\link[tidyr]{gather}} from the \pkg{tidyr} package for data.frames
}
\seealso{
\code{\link{gather.ExpressionSet}} \code{\link{gather.PharmacoSet}}
}

