% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_assay.PharmacoGx.R
\name{gather_assay.PharmacoGx}
\alias{gather_assay.PharmacoGx}
\title{gather_assay.PharmacoGx}
\usage{
gather_assay.PharmacoGx(x, sample_ids = NULL, gene_ids = NULL,
  data_type = "rna", sample_col = "cellid", gene_col = "Symbol")
}
\arguments{
\item{x}{A PharmacoGx object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene_ids}{A vector of gene ids.  Default is NULL (don't filter on gene id)}

\item{data_type}{Name of the component of the PharmacoGx object to gather}

\item{sample_col}{Name of the column in the pData data frame to use for filtering on sample id}

\item{gene_col}{Name of the column in the rowData data frame to use for filtering on gene id}
}
\value{
a tibble
}
\description{
gather_assay.PharmacoGx
}
\examples{
data('CCLEsmall', package='PharmacoGx')
gather_assay.PharmacoGx(CCLEsmall, sample_ids=c('143B', '23132-87'),
                         gene_ids=c('BRAF', 'EGFR'), data_type='rna')
set.seed(20)
gather_assay.PharmacoGx(CCLEsmall, sample_ids=sample(cellNames(CCLEsmall), 30),
                         gene_ids=c('BRAF', 'EGFR'), data_type='mutation')

gather_assay.PharmacoGx(CCLEsmall, sample_ids=sample(cellNames(CCLEsmall), 30),
 gene_ids=c('AK2', 'TFP1'), data_type='rnaseq', gene_col='gene_name')

}

