% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMultiAssay_generics.R
\name{gather_assay}
\alias{gather_assay}
\title{gather_assay}
\usage{
gather_assay(x, ...)
}
\arguments{
\item{x}{a PharmacoSet}

\item{...}{other arguments see \code{\link{gather_assay.PharmcoSet}}}
}
\description{
S3 generic for \code{\link{gather_assay.PharmcoSet}}
}
\seealso{
\code{\link{gather_assay.PharmcoSet}}
}

