% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_curvedata.PharmacoSet.R
\name{gather_curvedata.PharmacoSet}
\alias{gather_curvedata.PharmacoSet}
\title{Convert raw dose response curve data from a \linkS4class{PharmacoSet} into tidy format}
\usage{
\method{gather_curvedata}{PharmacoSet}(x, sample_ids = NULL,
  resp_ids = NULL)
}
\arguments{
\item{x}{PharmacoSet object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{resp_ids}{A vector of response ids.  Default is NULL (don't filter on response id)}

\item{resp_col}{Response variable to retrieve}
}
\value{
a data frame
}
\description{
Convert raw dose response curve data from a \linkS4class{PharmacoSet} into tidy format
}
\examples{
data('CCLEsmall', package='PharmacoGx')
gather_curvedata.PharmacoSet(CCLEsmall, sample_ids=c('CHL-1', 'SW1573'),
                         resp_ids=c('AEW541', 'Nilotinib'))
\dontrun{

#example to extract PARP inhibitor data from GDSC1000 dataset and what happens
#when more than one curve exists per compound/cell line combination

library(dplyr)
GDSC1000 <- PharmacoGx::downloadPSet('GDSC1000')

#get compounds that are PARP inhibitors
parp_inhibitors <- drugInfo(GDSC1000) \%>\%
        dplyr::filter(grepl('PARP', TARGET))

#get data for parp inhibitors
parpi_data <- gather_curvedata.PharmacoSet(GDSC1000, sample_ids=NULL,
                         resp_ids=parp_inhibitors$drugid)

#note warning about duplicate curves - this is because Olaparib was screened
#twice and has two drug ids - 1017 and 1495.  See parp_inhibtors dataframe

}

}

