% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMultiAssay_generics.R
\name{gather_curvedata}
\alias{gather_curvedata}
\title{gather_curvedata}
\usage{
gather_curvedata(x, ...)
}
\arguments{
\item{x}{a PharmacoSet}

\item{...}{other arguments see \code{\link{gather_curvedata.PharmcoSet}}}
}
\description{
S3 generic for \code{\link{gather_curvedata.PharmcoSet}}
}
\details{
gather_curvedata
}
\seealso{
\code{\link{gather_curvedata.PharmcoSet}}
}

