% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_response.data.frame.R
\name{gather_response.data.frame}
\alias{gather_response.data.frame}
\title{Convert response data from a data frame into tidy format}
\usage{
gather_response.data.frame(x, sample_ids = NULL, compound_ids = NULL,
  endpoints = NULL, resp_ids = NULL)
}
\arguments{
\item{x}{a data frame object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{compound_ids}{A vector of compound ids.  Default is NULL (don't filter on compound id)}

\item{endpoints}{A vector of endpoints.  Default is NULL (don't filter on endpoint)}

\item{resp_ids}{A vector of response ids.  Default is NULL (don't filter on response id)}
}
\value{
tibble
}
\description{
Convert response data from a data frame into tidy format
}
\examples{
data('dietlein_data')
gather_response.data.frame(dietlein_data, sample_ids=c('NCIH1703_LUNG', 'NCIH1792_LUNG'))
gather_response.data.frame(dietlein_data, compound_ids='KU60648')
gather_response.data.frame(dietlein_data, endpoints='pGI50')
gather_response.data.frame(dietlein_data, resp_ids='KU60648_pGI50')

}

