% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_response.PharmacoGx.R
\name{gather_response.PharmacoGx}
\alias{gather_response.PharmacoGx}
\title{gather_response.PharmacoGx}
\usage{
gather_response.PharmacoGx(x, sample_ids = NULL, resp_ids = NULL,
  resp_col = "ic50_published")
}
\arguments{
\item{x}{PharmacoGx object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{resp_ids}{A vector of response ids.  Default is NULL (don't filter on response id)}

\item{resp_col}{Response variable to retrieve}
}
\value{
tibble
}
\description{
gather_response.PharmacoGx
}
\examples{
data('CCLEsmall', package='PharmacoGx')
gather_response.PharmacoGx(CCLEsmall, sample_ids=c('CHL-1', 'SW1573'),
                         resp_ids=c('AEW541', 'Nilotinib'), resp_col='ic50_published')
}

