% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_response.PharmacoSet.R
\name{gather_response.PharmacoSet}
\alias{gather_response.PharmacoSet}
\title{Convert response data from a \linkS4class{PharmacoSet} into tidy format}
\usage{
gather_response.PharmacoSet(x, sample_ids = NULL, resp_ids = NULL,
  resp_col = "ic50_published")
}
\arguments{
\item{x}{PharmacoSet object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{resp_ids}{A vector of response ids.  Default is NULL (don't filter on response id)}

\item{resp_col}{Response variable to retrieve}
}
\value{
tibble
}
\description{
Convert response data from a \linkS4class{PharmacoSet} into tidy format
}
\examples{
data('CCLEsmall', package='PharmacoGx')
gather_response.PharmacoSet(CCLEsmall, sample_ids=c('CHL-1', 'SW1573'),
                         resp_ids=c('AEW541', 'Nilotinib'), resp_col='ic50_published')
}

