% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMultiAssay_generics.R
\name{gather_response}
\alias{gather_response}
\title{gather_response}
\usage{
gather_response(x, ...)
}
\arguments{
\item{x}{a data.frame or PharmacoSet}

\item{...}{other arguments see \code{\link{gather_response.PharmcoSet}} and \code{\link{gather_response.data.frame}}}
}
\description{
S3 generic for \code{\link{gather_response.PharmcoSet}} and \code{\link{gather_response.data.frame}}
}
\details{
gather_response
}
\seealso{
\code{\link{gather_response.PharmcoSet}} \code{\link{gather_response.data.frame}}
}

