% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_genetic_vs_genetic_df.data.frame.R
\name{make_genetic_vs_genetic_df.data.frame}
\alias{make_genetic_vs_genetic_df.data.frame}
\title{Create a genetic_vs_genetic data frame from two tall data frames}
\usage{
make_genetic_vs_genetic_df.data.frame(df, df2 = NULL, sample_ids = NULL,
  gene1 = NULL, gene2 = NULL, data_type1 = NULL, data_type2 = NULL)
}
\arguments{
\item{df}{data.frame object}

\item{df2}{Second data.frame object.  Default is NULL (just use the first data.frame)}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene1}{First gene(s)}

\item{gene2}{Second gene(s)}

\item{data_type1}{Data type for gene 1}

\item{data_type2}{Data type for gene 2}

\item{sample_col}{Name of the column in the pData data frame to use for filtering on sample id}

\item{gene_col}{Name of the column in the rowData data frame to use for filtering on gene id}
}
\value{
tibble
}
\description{
Create a genetic_vs_genetic data frame from two tall data frames
}
\examples{
data('CCLEsmall', package='PharmacoGx')

dat1 <- gather_assay.PharmacoSet(CCLEsmall, sample_ids=PharmacoGx::cellNames(CCLEsmall),
  gene_ids = c('RBM5', 'LAP3', 'CFTR'), data_type = 'rna')
dat2 <- gather_assay.PharmacoSet(CCLEsmall, sample_ids=PharmacoGx::cellNames(CCLEsmall),
  gene_ids = c('RBM5', 'LAP3', 'CFTR'), data_type = 'rnaseq', gene_col='gene_name')

gvg_df <- make_genetic_vs_genetic_df.data.frame(df=dat1, df2=dat2, sample_ids=PharmacoGx::cellNames(CCLEsmall), gene1='RBM5', gene2='RBM5',
 data_type1='rna', data_type2='rnaseq')
gvg_df
}

