% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_genetic_vs_genetic_df.PharmacoSet.R
\name{make_genetic_vs_genetic_df.PharmacoSet}
\alias{make_genetic_vs_genetic_df.PharmacoSet}
\title{Create a genetic_vs_genetic data frame from the data in a PharmacoSet}
\usage{
\method{make_genetic_vs_genetic_df}{PharmacoSet}(x, sample_ids = NULL, gene1,
  gene2, data_type1, data_type2, sample_col = "cellid",
  gene_col1 = "Symbol", gene_col2 = "Symbol")
}
\arguments{
\item{x}{PharmacoSet object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene1}{First gene(s)}

\item{gene2}{Second gene(s)}

\item{data_type1}{Data type for gene 1}

\item{data_type2}{Data type for gene 2}

\item{sample_col}{Name of the column in the pData data frame to use for filtering on sample id}

\item{gene_col}{Name of the column in the rowData data frame to use for filtering on gene id}
}
\value{
a data frame in gvg_df format
}
\description{
Create a genetic_vs_genetic data frame from the data in a \linkS4class{PharmacoSet}
}
\examples{
data('CCLEsmall', package='PharmacoGx')

gvg_df <- make_genetic_vs_genetic_df.PharmacoSet(CCLEsmall, sample_ids=PharmacoGx::cellNames(CCLEsmall), gene1='RBM5', gene2='RBM5',
 data_type1='rna', data_type2='rnaseq', gene_col1 = "Symbol", gene_col2 = "gene_name")
gvg_df
}

