% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMultiAssay_generics.R
\name{make_genetic_vs_genetic_df}
\alias{make_genetic_vs_genetic_df}
\title{make_genetic_vs_genetic_df}
\usage{
make_genetic_vs_genetic_df(x, ...)
}
\arguments{
\item{x}{a data.frame or PharmacoSet}

\item{...}{other arguments see \code{\link{make_genetic_vs_genetic_df.PharmcoSet}} and \code{\link{make_genetic_vs_genetic_df.data.frame}}}
}
\description{
S3 generic for \code{\link{make_genetic_vs_genetic_df.PharmcoSet}} and \code{\link{make_genetic_vs_genetic_df.data.frame}}
}
\seealso{
\code{\link{make_genetic_vs_genetic_df.PharmcoSet}} \code{\link{make_genetic_vs_genetic_df.data.frame}}
}

