% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_response_vs_genetic_df.data.frame.R
\name{make_response_vs_genetic_df.data.frame}
\alias{make_response_vs_genetic_df.data.frame}
\title{Create a response_vs_genetic data frame from two tall data frames}
\usage{
\method{make_response_vs_genetic_df}{data.frame}(df, df2 = NULL,
  sample_ids = NULL, gene_ids = NULL, data_types = NULL,
  compound_ids = NULL, endpoints = NULL, resp_ids = NULL)
}
\arguments{
\item{df}{a data frame in tall_df format containing the genetic data (and response data if df2 not set)}

\item{df2}{a data frame in either tall_df or resp_df format containing the response data.
Default is NULL (just use df)}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene_ids}{Gene ids in df.  Default is NULL (don't filter on gene id)}

\item{data_types}{Data type for genetic data.  Default is NULL (don't filter on data type)}

\item{compound_ids}{A vector of compound ids.  Default is NULL (don't filter on compound id)}

\item{endpoints}{A vector of endpoints.  Default is NULL (don't filter on endpoint)}

\item{resp_ids}{A vector of response ids.  Default is NULL (don't filter on response id)}
}
\value{
a data frame in rvg_df format
}
\description{
Create a response_vs_genetic data frame from two tall data frames
}
\examples{
data('CCLEsmall', package='PharmacoGx')

genetic_data <- gather_assay.PharmacoSet(CCLEsmall, sample_ids=PharmacoGx::cellNames(CCLEsmall),
  gene_ids = c('RBM5', 'LAP3', 'CFTR'), data_type = 'rna')
resp_data <- gather_response.PharmacoSet(CCLEsmall, sample_ids=PharmacoGx::cellNames(CCLEsmall),
                         resp_ids=c('AEW541', 'Nilotinib'), resp_col='ic50_published')

rvg_df <- make_response_vs_genetic_df.data.frame(df=genetic_data, df2=resp_data, sample_ids=PharmacoGx::cellNames(CCLEsmall), gene_ids='RBM5',
 data_types='rna', resp_ids='Nilotinib')
rvg_df
}

