% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneticVsGeneticTibble.PharmacoGx.R
\name{makeGeneticVsGeneticTibble.PharmacoGx}
\alias{makeGeneticVsGeneticTibble.PharmacoGx}
\title{makeGeneticVsGeneticTibble.PharmacoGx}
\usage{
makeGeneticVsGeneticTibble.PharmacoGx(x, sample_ids = NULL, gene1, gene2,
  data_type1, data_type2, sample_col = "cellid", gene_col1 = "Symbol",
  gene_col2 = "Symbol")
}
\arguments{
\item{x}{PharmacoGx object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene1}{First gene}

\item{gene2}{Second gene}

\item{data_type1}{Data type for gene 1}

\item{data_type2}{Data type for gene 2}

\item{sample_col}{Name of the column in the pData data frame to use for filtering on sample id}

\item{gene_col}{Name of the column in the rowData data frame to use for filtering on gene id}
}
\value{
tibble
}
\description{
makeGeneticVsGeneticTibble.PharmacoGx
}
\examples{
data('CCLEsmall', package='PharmacoGx')

gvg_df <- makeGeneticVsGeneticTibble.PharmacoGx(CCLEsmall, sample_ids=cellNames(CCLEsmall), gene1='RBM5', gene2='RBM5',
 data_type1='rna', data_type2='rnaseq', gene_col1 = "Symbol", gene_col2 = "gene_name")
gvg_df
CancerCellLines::plotGeneticVsGeneticPoint(gvg_df)
}

