% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneticVsGeneticTibble.PharmacoSet.R
\name{makeGeneticVsGeneticTibble.PharmacoSet}
\alias{makeGeneticVsGeneticTibble.PharmacoSet}
\title{makeGeneticVsGeneticTibble.PharmacoSet}
\usage{
makeGeneticVsGeneticTibble.PharmacoSet(x, sample_ids = NULL, gene1, gene2,
  data_type1, data_type2, sample_col = "cellid", gene_col1 = "Symbol",
  gene_col2 = "Symbol")
}
\arguments{
\item{x}{PharmacoSet object}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene1}{First gene(s)}

\item{gene2}{Second gene(s)}

\item{data_type1}{Data type for gene 1}

\item{data_type2}{Data type for gene 2}

\item{sample_col}{Name of the column in the pData data frame to use for filtering on sample id}

\item{gene_col}{Name of the column in the rowData data frame to use for filtering on gene id}
}
\value{
tibble
}
\description{
makeGeneticVsGeneticTibble.PharmacoSet
}
\examples{
data('CCLEsmall', package='PharmacoGx')

gvg_df <- makeGeneticVsGeneticTibble.PharmacoSet(CCLEsmall, sample_ids=cellNames(CCLEsmall), gene1='RBM5', gene2='RBM5',
 data_type1='rna', data_type2='rnaseq', gene_col1 = "Symbol", gene_col2 = "gene_name")
gvg_df
CancerCellLines::plotGeneticVsGeneticPoint(gvg_df)
}

