% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRespVsGeneticTibble.data.frame.R
\name{makeRespVsGeneticTibble.data.frame}
\alias{makeRespVsGeneticTibble.data.frame}
\title{makeRespVsGeneticTibble.data.frame}
\usage{
makeRespVsGeneticTibble.data.frame(df, df2 = NULL, sample_ids = NULL,
  gene_ids = NULL, data_types = NULL, compound_ids = NULL,
  endpoints = NULL, resp_ids = NULL)
}
\arguments{
\item{df}{data.frame object}

\item{df2}{Second data.frame object.  Default is NULL (just use the first data.frame)}

\item{sample_ids}{A vector of sample ids.  Default is NULL (don't filter on sample id)}

\item{gene_ids}{Gene ids in df.  Default is NULL (don't filter on gene id)}

\item{data_types}{Data type for genetic data.  Default is NULL (don't filter on data type)}

\item{compound_ids}{A vector of compound ids.  Default is NULL (don't filter on compound id)}

\item{endpoints}{A vector of endpoints.  Default is NULL (don't filter on endpoint)}

\item{resp_ids}{A vector of response ids.  Default is NULL (don't filter on response id)}
}
\value{
tibble
}
\description{
makeRespVsGeneticTibble.data.frame
}
\examples{
data('CCLEsmall', package='PharmacoGx')

genetic_data <- gather_assay.PharmacoSet(CCLEsmall, sample_ids=PharmacoGx::cellNames(CCLEsmall),
  gene_ids = c('RBM5', 'LAP3', 'CFTR'), data_type = 'rna')
resp_data <- gather_response.PharmacoSet(CCLEsmall, sample_ids=PharmacoGx::cellNames(CCLEsmall),
                         resp_ids=c('AEW541', 'Nilotinib'), resp_col='ic50_published')

rvg_df <- makeRespVsGeneticTibble.data.frame(df=genetic_data, df2=resp_data, sample_ids=PharmacoGx::cellNames(CCLEsmall), gene_ids='RBM5',
 data_types='rna', resp_ids='Nilotinib')
rvg_df
CancerCellLines::plotGeneticVsGeneticPoint(gvg_df)
}

