% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df_format.R
\name{check_df_format}
\alias{check_df_format}
\title{Checks the format of a data frame}
\usage{
check_df_format(df, format, dev_mode = FALSE)
}
\arguments{
\item{df}{query data frame}

\item{format}{format to test - see ?data_formats}

\item{dev_mode}{stop if test is false.  Default is FALSE}
}
\value{
string of data frame type
}
\description{
Helper function to determine whether a supplied data frame is a specific fomat
defined in the package.  Either returns a boolean or throws a stop.
}
\examples{
data(dietlein_data)
biotidyr:::check_df_format(dietlein_data, 'resp_df')

data(example_tall_df)
biotidyr:::check_df_format(example_tall_df, 'tall_df')

data(mtcars)
biotidyr:::check_df_format(mtcars, 'tall_df')
}

