% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crush.R
\name{crush}
\alias{crush}
\title{Crush cell line id format}
\usage{
crush(x, upper_case = FALSE)
}
\arguments{
\item{x}{character vector of ids}

\item{upper_case}{boolean - return in upper case or not.  Default FALSE}
}
\value{
character vector of crushed ids
}
\description{
Helper function to remove punctuation and whitespace from cell line id's.  This helps convert id's between datasets.
}
\examples{

data("CCLEsmall", package='PharmacoGx')
PharmacoGx::cellNames(CCLEsmall)[1:20]
crush(PharmacoGx::cellNames(CCLEsmall)[1:20])
}

